Feature/clean up generator (#18)
* Moved prompts to a different file

* More todo

* Ordered to do into difficulty

* Added todo

* Wrote read me and fixed var name

* Added goodbuy and wip cleanup of app.js

* Using yo lifcycle and condensed as much as possible
