+++ README.md
@@ -1,10 +1,32 @@
 [![Build Status](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule.svg?branch=master)](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule)
 [![Coverage Status](https://coveralls.io/repos/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg?branch=master)](https://coveralls.io/github/Aaron-K-T-Berry/generator-ansible-molecule?branch=master)
 [![Known Vulnerabilities](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg)](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule)
-# generator-ansible-molecule
+# Generator-Ansible-Molecule
+Yeoman generator to create ansible roles with molecule testing framework including optional sample ci cd configs.
+
+### Installation
+``` 
+npm i generator-ansible-molecule
+```
+
+### Running
+```
+yo ansible-molecule
+```
+
+And follow the prompts to create a new ansible role.
+
+### Support
+Currently this generator is being tested against the following node versions.
+- Node v11
+- Node v10
+- Node v8
+
+### Contributing
+Feel free to post issues, and create prs with an changes, fixes or enhancements.
 
 <!-- TODO [OPTS] Add option to specify path of install -->
 <!-- TODO [OPTS] Print out all options after yo message -->
-<!-- TODO [CICD] Add more ci cd options aim for all free ones-->
+<!-- TODO [CICD] Add more ci cd options aim for all free ones on github marketplace -->
 <!-- TODO [TESTS] Split up tests more logically -->
-<!-- TODO [CLEANUP] Refactor generator to be more compact -->
+<!-- TODO [REPO] Add semantic release to ci pipeline -->
